+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Changed
+ - Hex color field behavior
+
 ## 2.7.2.1 - 2025-08-31
 ### Added
  - Custom version compatibility checker